Rename glvd to glvd-data
We want to provide some client tooling from the same source.  So make
sure we have the name glvd free for that use.
