+++ pyproject.toml
@@ -11,7 +11,7 @@ packages = [
 ]
 
 [tool.poetry.scripts]
-glvd = 'glvd.cli.__main__:main'
+glvd-data = 'glvd.cli.data.__main__:main'
 
 [tool.poetry.dependencies]
 python = ">=3.11"
