+++ src/glvd/cli/data/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: MIT
+
+from __future__ import annotations
+
+from ..registry import CliRegistry
+
+
+cli = CliRegistry()
