+++ src/glvd/cli/data/ingest_debsrc.py
@@ -14,9 +14,9 @@ from sqlalchemy.ext.asyncio import (
     create_async_engine,
 )
 
-from ..database import Base, DistCpe, Debsrc
-from ..data.debsrc import DebsrcFile
-from ..data.dist_cpe import DistCpeMapper
+from glvd.database import Base, DistCpe, Debsrc
+from glvd.data.debsrc import DebsrcFile
+from glvd.data.dist_cpe import DistCpeMapper
 from . import cli
